% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gradient.R
\name{make_gradient}
\alias{make_gradient}
\title{Make Gradient}
\usage{
make_gradient(string_lab = NULL, colors = NULL)
}
\arguments{
\item{string_lab}{character(1) string, a text label to which the gradient colors will be applied to}

\item{colors}{character(...) string, a vector of at least two characters representing valid color names or hex codes,uses linear interpolation to create gradient color scheme}
}
\value{
make_lab, the text label with html span tags that apply the gradient color scheme specified in the colors argument
}
\description{
'This function takes a string label and a vector of at least two colors as an argument and applies gradient colors to the text. The function must be used with either geom_richtext, geom_textbox, render_markdown or render_textbox functions from ggtext.
}
